==> misc_files/Spy-Plane-Project/2-SpyPlane-RandomForestClassifier-Unsolved.ipynb <==
# Random Forest Model

This notebook pulls in the cleaned data from the data_cleaning.ipynb notebook and uses that to create and evaluate
a random forest model.

This projects is based off the Buzzfeed news article on identifying spy planes found [here](https://www.buzzfeednews.com/article/peteraldhous/hidden-spy-planes), using the data and code adapted from their github repository [here](https://github.com/BuzzFeedNews/2017-08-spy-plane-finder).

## Instructions

Follow the directions in any cell that does not contain code. If a cell does contain code, run this before moving on to the next cell

%matplotlib inline
#import packages
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

#sci-kit learn is a library with machine learning algorithms
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix

# This relies on output from a previous notebook!
# If this cell does not work, try using the pregenerated data instead
#planes_labeled = pd.read_csv("/mnt/data/spyplane-data/pregenerated_planes_labeled.csv")
planes_labeled = pd.read_csv("../data/spyplane-data/planes_labeled.csv")
planes_labeled.head()

### Background 

The random forest classifier works by constructing a multitude of decision trees at training time and outputting the class that is the mode of the classes or mean prediction of the individual trees. You can find the random forest classifier documentation for sci-kit learn [here](https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.RandomForestClassifier.html).

### Format data

We'll first format our data for the model. 

#factorize the classes
y = pd.factorize(planes_labeled['class'])[0]
y[0:5]

#create labels
labels = pd.factorize(planes_labeled['class'])[1]
labels

#create X and drop columns that won't be used for training
X = planes_labeled.drop(['adshex','class', 'type'], axis = 1)
X.head()

### Train and test data sets

We want to be sure to set aside some data to test this model so that we can determine it's accuracy. Here, we'll use sci-kit learn to split our data into these groups. We'll set a test size of 30% and use the remaining 70% to train on. 

#split the data into test and train sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2)

print(X_train.shape, X_test.shape)

### Train model

Now we'll train our model and make predictions. We'll seed the random number generator so that we get the same values.

#set a seed so that the results are reproducible
np.random.seed(415)

#create and train model using RandomForestClassifier()

#predict the classes

#print first 10 predictions

#visualize the number of classified surveillence and other planes in the test set

cmtx = pd.DataFrame(
    confusion_matrix(y_test, predictions), 
    index=['true:other', 'true:surveillence'], 
    columns=['predicted:other', 'predicted:surveillence']
)

cmtx

#look at the overall accuracy from the numbers in the table above
calculated_accuracy =
calculated_accuracy

#calculate the probabilities that each plane is a surviellance plane
predict_prob = spy_model.predict_proba(X_test)
predict_prob[0:10]
