==> misc_files/Spy-Plane-Project/4-SpyPlane-OptimizingModels-Unsolved.ipynb <==
# Optimizing Models

This notebook pulls in the model from the SpyPlane-RandomForestClassifier.ipynb notebook and optimizes the random forest model created there.

This projects is based off the Buzzfeed news article on identifying spy planes found [here](https://www.buzzfeednews.com/article/peteraldhous/hidden-spy-planes), using the data and code adapted from their github repository [here](https://github.com/BuzzFeedNews/2017-08-spy-plane-finder).

## Instructions

Follow the directions in any cell that does not contain code. If a cell does contain code, run this before moving on to the next cell

%matplotlib inline
#import packages
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

#sci-kit learn is a library with machine learning algorithms
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn import metrics
from sklearn.model_selection import RandomizedSearchCV, GridSearchCV

### Load Data and Train Model

Repeat the data formatting steps from the previous notebooks.

# This relies on output from a previous notebook!
# If this cell does not work, try using the pregenerated data instead
#planes_labeled = pd.read_csv("/mnt/data/spyplane-data/pregenerated_planes_labeled.csv")
planes_labeled = pd.read_csv("/mnt/data/spyplane-data/planes_labeled.csv")

#format data by removing non-numeric columnns and factorize the class
X = planes_labeled.drop(['adshex','class', 'type'], axis = 1)
y = pd.factorize(planes_labeled['class'])[0]

#split the data into train and test sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.3, random_state=42)

## Hyperparameters

Hyperparameters can be thought of as the settings of an algorithm that can be adjusted to optimize performance. These are something you will set before training the model. You are unlikely to know what the best settings for these will be ahead of time, and tuning a model is where machine learning turns from a science into trial-and-error based engineering.

In the case of a random forest, hyperparameters include the number of decision trees in the forest and the number of features considered by each tree when splitting a node. Scikit-Learn implements default hyperparameters for all models, but these are not guaranteed to be optimal for a problem. 

Below, we can see the default parameters that were used in our above model:

We can see the effects of altering hyperparameters by increasing the number of estimators (a.k.a the number of decision trees) on the compute time, precision, and recall:

%%time
#try model with different parameter for number of trees (default of 10)
np.random.seed(415)
spy_model = RandomForestClassifier(n_estimators = 10)
spy_model.fit(X_train,y_train)

#calculate predictions on test set
predictions = spy_model.predict(X_test)

#metrics
precision = metrics.precision_score(y_true = y_test, y_pred = predictions)
recall = metrics.recall_score(y_true = y_test, y_pred = predictions)
accuracy = metrics.accuracy_score(y_true = y_test, y_pred = predictions)
print('Precision: ' + str(precision) + ' Recall: ' + str(recall) + 
      ' Accuracy: ' + str(accuracy))

%%time
#try model with different parameter for number of trees (100)
np.random.seed(415)
model_higher_n = RandomForestClassifier(n_estimators = 100)
model_higher_n.fit(X_train,y_train)

#calculate predictions on test set
predictions_higher_n = model_higher_n.predict(X_test)

#metrics
precision_higher_n = metrics.precision_score(y_true = y_test, y_pred = predictions_higher_n)
recall_higher_n = metrics.recall_score(y_true = y_test, y_pred = predictions_higher_n)
accuracy_higher_n = metrics.accuracy_score(y_true = y_test, y_pred = predictions_higher_n)
print('Precision: ' + str(precision_higher_n) + ' Recall: ' + str(recall_higher_n) + 
      ' Accuracy: ' + str(accuracy_higher_n))

spy_model = RandomForestClassifier(n_estimators = 10)
print('Parameters currently in use:\n')
print(spy_model.get_params())

Here are a few commonly adjusted hyperparameters of random forests (See this [article](https://medium.com/@ODSC/optimizing-hyperparameters-for-random-forest-algorithms-in-scikit-learn-d60b7aa07ead) for more detail):

- `n_estimators`: Random forest models are ensembles of decision trees. N_estimators controls the number of decision trees used in the model.
     
     
- `max_features`: Random forest models randomly resample features prior to determining the best split. This determines the number of features to resample. 


- `max_depth`: Each tree in the random forest model makes multiple splits to isolate homogeneous groups of outcomes. Max_depth determines the max number of levels. 


- `min_samples_split`: Controls the minimum number of samples required to split each node. 


- `min_samples_leaf`: Much like stopping the growth of trees once a minimum number of samples per split is reached, we can set the minimum number of samples required for each leaf. 
    
    
- `bootstrap`: method for sampling data points (with or without replacement)


- `class_weight`: Weights associated with classes. The “balanced” mode uses the values of y to automatically adjust weights inversely proportional to class frequencies

# Activity - Hyperparameter optimization

Pick a metric and see if you can find the hyperparameters that optimize this. Feel free to use google to research the above hyperparameters as well as others that are used in Random Forest Classifiers. Official documentation can be found [here](https://scikit-learn.org/stable/modules/generated/sklearn.ensemble.RandomForestClassifier.html).

**Spoiler Alert: Don't scroll below this section in the notebook!**

%%time

########################################################
#fill in the hyperparameter value(s) you'd like to try - feel free to add others
n_estimators = 5
max_depth = 5
max_features = 'sqrt'
min_samples_split = 10
bootstrap = False
class_weight = 'none'
########################################################

#create and fit model with your chosen hyperparameter values
np.random.seed(415)
your_model = RandomForestClassifier(n_estimators=n_estimators, max_depth=max_depth, max_features=max_features, 
                               min_samples_split=min_samples_split, bootstrap=bootstrap)
your_model.fit(X_train,y_train)

#calculate predictions on test set
your_predictions = your_model.predict(X_test)

#metrics
your_precision = metrics.precision_score(y_true = y_test, y_pred = your_predictions)
your_recall = metrics.recall_score(y_true = y_test, y_pred = your_predictions)
your_accuracy = metrics.accuracy_score(y_true = y_test, y_pred = your_predictions)
your_f1_score = metrics.f1_score(y_true = y_test, y_pred = your_predictions)

print('Accuracy: ' + str(round(your_accuracy,4)) + '\nPrecision: ' + str(round(your_precision,4)) + ', Recall: ' + str(round(your_recall,4)) + '\n\nF1 Score: '  + str(round(your_f1_score,6)) + '\n')

### Submit Result to Leaderboard

#### When you are happy with your model, run the following cell to submit your result to the leaderboard!

import os
import csv
import time

import boto3
from botocore import UNSIGNED
from botocore.client import Config

user = os.environ['DOMINO_STARTING_USERNAME']

score_dict = {'user':[user,],
        'f1':[your_f1_score,], 'precision':[your_precision,], 'recall':[your_recall,],
        'n_estimators':[n_estimators,], 'max_depth':[max_depth,], 'max_features':[max_features,], 
        'min_samples_split':[min_samples_split,], 'bootstrap':[bootstrap,], 'class_weight':[class_weight,]
       }

filename = user + '.csv'
pd.DataFrame(score_dict, index=[0,]).to_csv(filename, index=False)
    
time.sleep(0.5)

client = boto3.client('s3', aws_access_key_id='', aws_secret_access_key='')
client._request_signer.sign = (lambda *args, **kwargs: None)

client.upload_file(user + ".csv", "workshop-leaderboard", "lm-workshop/" + user + ".csv")

print("Data upload succesfully")
pd.read_csv(filename).head()

#### Spoilers below
.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

.

### Random Search Cross Validation

To check different values for these hyperparameters, we can use scikit-learn’s “RandomizedSearchCV” to define a grid of hyperparameter ranges and randomly sample from the grid, performing cross validation with each combination of values.

# Number of trees in random forest
n_estimators = [int(x) for x in np.linspace(start = 1000, stop = 3000, num = 200)]
# Number of features to consider at every split
max_features = ['auto', 'sqrt', 'log2']
# Maximum number of levels in tree
max_depth = [int(x) for x in np.linspace(10, 110, num = 11)]
max_depth.append(None)
# Minimum number of samples required to split a node
min_samples_split = [2, 5, 10]
# Minimum number of samples required at each leaf node
min_samples_leaf = [1, 2, 4]
# Method of selecting samples for training each tree
bootstrap = [True, False]
#Weights associated with classes, The “balanced” mode uses the values of y to automatically adjust weights 
#inversely proportional to class frequencies
class_weight=['balanced',None]
# Create the random grid

random_grid = {'n_estimators': n_estimators,
               'max_features': max_features,
               'max_depth': max_depth,
               'min_samples_split': min_samples_split,
               'min_samples_leaf': min_samples_leaf,
               'bootstrap': bootstrap,
               'class_weight':class_weight}

# Use the random grid to search for best hyperparameters
# First create the base model to tune
np.random.seed(415)
model = RandomForestClassifier()
# Random search of parameters, using 3 fold cross validation, 
# search across 100 different combinations, and use all available cores
model_rgs = RandomizedSearchCV(estimator = model, param_distributions = random_grid, scoring = 'recall',
                               n_iter = 15, cv = 3, verbose=2, random_state=42, n_jobs = -1)
# Fit the random search model
model_rgs.fit(X_train, y_train)

#look at randomized search best parameters
model_rgs.best_params_

# Create a model based on the hyperparameters from the above grid search
np.random.seed(415)
model_tuned = RandomForestClassifier(**model_rgs.best_params_)
model_tuned.fit(X_train, y_train)

predictions_tuned = model_tuned.predict(X_test)

#metrics
precision_rgs = metrics.precision_score(y_true = y_test, y_pred = predictions_tuned)
recall_rgs = metrics.recall_score(y_true = y_test, y_pred = predictions_tuned)
accuracy_rgs = metrics.accuracy_score(y_true = y_test, y_pred = predictions_tuned)
print('Randomized Grid Search Metrics: Precision: ' + str(precision_rgs) + ' Recall: ' + str(recall_rgs) + 
      ' Accuracy: ' + str(accuracy_rgs))

### Feature Engineering

We'll look at the importance of different features in the data set on our predictions and make a visualization of these.

#make lists of the features (column names) and the feature importances returned from our model
features = X.columns
importances = list(model_tuned.feature_importances_)

#combine the names and importances and sort
feature_importance = list(zip(features, np.round(importances,2)))
feature_importance.sort(key=lambda x:x[1], reverse = True)
feature_importance

#plot the feature importances
features_ranked = list(zip(*feature_importance))[0]
feat_imp_ranked = list(zip(*feature_importance))[1]

fig = plt.figure(figsize = [15,5])
plt.xticks(fontsize = 14, rotation=75)
plt.yticks(fontsize = 14)
plt.xlabel('Plane Feature', fontsize = 18)
plt.ylabel('Feature Importance', fontsize = 18)
plt.bar(features_ranked, feat_imp_ranked, align='center');

Sci-kit learn's feature importance uses `MeanDecreaseGini`, which measures the extent to which each variable plays a role in partitioning the data into the defined classes.

As seen in the news article analysis, the `steer1` and `steer2` variables, quantifying the frequency of turning hard to the left, were the most important to the model.

# Activity 

Try training a model with a limited amout of features and see how the metrics change:

#add different columns based on feature importance here

X_limited = planes_labeled[['steer1', ]] #add other features here

#factorize the class
y = pd.factorize(planes_labeled['class'])[0]

#split the data into train and test sets
X_train_lim, X_test_lim, y_train, y_test = train_test_split(X_limited, y, test_size=0.3, random_state=42)

# Create a model based on the hyperparameters from the above grid search
np.random.seed(415)
model_tuned = RandomForestClassifier(n_estimators=1100, max_depth=50, max_features='sqrt', 
                                     min_samples_split=4, bootstrap=False, class_weight='balanced')

#train model with only features above
model_tuned.fit(X_train_lim, y_train)

#predictions
predictions_tuned = model_tuned.predict(X_test_lim)

#metrics
precision_tuned = metrics.precision_score(y_true = y_test, y_pred = predictions_tuned)
recall_tuned = metrics.recall_score(y_true = y_test, y_pred = predictions_tuned)
accuracy_tuned = metrics.accuracy_score(y_true = y_test, y_pred = predictions_tuned)
print('Randomized Grid Search Metrics: Precision: ' + str(precision_tuned) + ' Recall: ' + str(recall_tuned) + 
      ' Accuracy: ' + str(accuracy_tuned))
