==> misc_files/Spy-Plane-Project/5-SpyPlane-CreateModel.ipynb <==
# Create and Use model

This notebook pulls in the model from the SpyPlane-OptimizingModel.ipynb notebook and uses that model to train the remaining data

This projects is based off the Buzzfeed news article on identifying spy planes found [here](https://www.buzzfeednews.com/article/peteraldhous/hidden-spy-planes), using the data and code adapted from their github repository [here](https://github.com/BuzzFeedNews/2017-08-spy-plane-finder).

%matplotlib inline
#import packages
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

#sci-kit learn is a library with machine learning algorithms
from sklearn.ensemble import RandomForestClassifier

#package for saving our ML model
import pickle

# This relies on output from a previous notebook!
# If this cell does not work, try using the pregenerated data instead
#planes_labeled = pd.read_csv("/mnt/data/spyplane-data/pregenerated_planes_labeled.csv")
planes_labeled = pd.read_csv("/mnt/data/spyplane-data/planes_labeled.csv")

#format data by removing non-numeric columnns and factorize the class
X = planes_labeled[['steer1', 'steer2', 'steer4', 'steer5', 'steer6', 'squawk_1', 'altitude3']]
y = pd.factorize(planes_labeled['class'])[0]

# Create a model (based on the best hyperparameters from the previous notebook)
np.random.seed(415)
model_tuned = RandomForestClassifier(
    n_estimators=1100,
    max_depth=40,
    max_features='auto', 
    min_samples_split=10,
    bootstrap=False
)

#train model with only features above
model_tuned.fit(X, y)

### Save the model to disk
This is often referred to as **serializing** the model. Notice that after this step, we will not reference the variable `model_tuned` for the rest of the notebook! Instead, a later cell will show how to load the model from this file. This mimics the way a model will often have prediction code running separately from the training code.

# save the model to disk
pickle.dump(model_tuned, open('/mnt/data/spyplane-data/SpyPlane_RandomForest.sav', 'wb'))

### Classify all data

#### Gather and format data
First we'll remove all of the training data and the known federal planes from the entire data set (which is in the planes_features file.

#read in all data
feds = pd.read_csv("/mnt/data/spyplane-data/feds_adshex_list.csv")
train = pd.read_csv("/mnt/data/spyplane-data/train.csv")
planes = pd.read_csv('/mnt/data/spyplane-data/planes_features.csv')

#first gather list of federal plane identifiers to remove
fed_ids = list(feds['adshex'])
len(fed_ids)

#next gather list of training list identifiers to remove 
remove = fed_ids + list(train['adshex'])
len(remove)

classify = planes[~planes['adshex'].isin(remove)] 

#look at number of rows and columns in the raw data 
classify.shape

X_all = classify[['steer1', 'steer2', 'steer4', 'steer5', 'steer6', 'squawk_1', 'altitude3']]

X_all.head()

#### Read model from disk
The is often referred to as **deserializing** the model, and requires the file where we saved the model to disk earlier.

model_deserialized = pickle.load(open('/mnt/data/spyplane-data/SpyPlane_RandomForest.sav', 'rb'))

#### Make Predictions

Then we'll use our deserialized model to output those planes that were determined to be potential surveillance plans and join this data with the [FAA aircraft registration database](https://www.faa.gov/licenses_certificates/aircraft_certification/aircraft_registry/releasable_aircraft_download/), which gives the planes’ registration numbers and the organizations they are registered to.|

real_predictions = model_deserialized.predict(X_all)

#look at number of predicted spy planes
sum(real_predictions)

#create data frame with only those potential spy planes
candidates = classify[real_predictions == 1]

candidates.head()

#read in FAA data
faa = pd.read_csv("/mnt/data/large-spyplane-data/faa-registration.csv")
faa.head()

#look at the column names in the FAA registration dataframe
faa.columns

#seperate out the columns we want to use
plane_info = faa[['N-NUMBER', 'NAME', 'MODE S CODE HEX']].copy()
plane_info.rename(columns = {'N-NUMBER':'n_number', 'NAME':'name', 'MODE S CODE HEX':'adshex'}, inplace = True) 

spy_candidates = candidates.merge(plane_info, on = 'adshex', how = 'left')

#### Look at predicted probabilities

Here, we'll calculate the probabilities and sort them in descending order.

#get list of the probabilities
probability_pred = model_deserialized.predict_proba(X_all)

#add the surveillance plane classifications to the data frame and sort
classify_prob = classify.copy()
classify_prob.loc[:,'spy_prob'] = probability_pred[:,1]
classify_prob.sort_values(by = 'spy_prob', ascending=False, inplace=True)

classify_prob.head()

#merge with FAA names and registration numbers
classify_prob_faa= classify_prob.merge(plane_info, on = 'adshex', how = 'left')

#seperate out only those rows with probabilites greater than 0.5 and the relevant columns
relevant_cols = ['adshex', 'type', 'spy_prob', 'n_number', 'name', 'squawk_1', 'steer1', 'steer2', 'steer4', 'steer5', 'steer6', 'altitude3']
candidates_with_prob = classify_prob_faa.loc[classify_prob_faa['spy_prob'] > 0.5, relevant_cols]

#look at the top 15 results
candidates_with_prob.head(15)

#save the spy candidates data frame to a csv file
candidates_with_prob.to_csv("/mnt/data/spyplane-data/spy_candidates.csv", index = False)

#save the confirmed federal surveillance planes with their relevant data to file
feds_data = planes_labeled[planes_labeled['adshex'].isin(fed_ids)]
feds_data = feds_data[['adshex', 'steer1', 'steer2', 'steer4', 'steer5', 'steer6', 'squawk_1', 'altitude3']]
feds_data.to_csv('/mnt/data/spyplane-data/feds_data.csv')
